requirements.txt
pytest==8.3.2
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
allure-pytest==2.13.5
minio~=7.2.5
pytest-dotenv